Escape isBanned throwing from API - adds plugin for compatibility - rename login event
